create an order

+++ foundation/Foundation2/product.cs
@@ -3,17 +3,17 @@ using System;
 public class Product{
     private string _name;
     private int _productId;
-    private float _price;
+    private double _price;
     private int _quantity;
 
-    public Product(string name, int productId, float price, int quantity){
+    public Product(string name, int productId, double price, int quantity){
         _name = name;
         _productId = productId;
         _price = price;
         _quantity = quantity;
     }
 
-    public float GetTotalCost(){
+    public double GetTotalCost(){
         return _price * _quantity;
     }
